pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embeddings-viz"
version = "0.1.0"
description = "Embedding projector — visualize how words relate in semantic space"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "numpy",
    "pyyaml",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.0.0",
    "torch",
    "tqdm",
    "transformers",
    "umap-learn",
]

[project.scripts]
embeddings-viz = "embeddings_viz.app:main"
